✨ Move pusher to browser client
